requirements.txt
Flask==2.1.0
Pillow==8.4.0
